+++ spec/examples/disallow_nil_on_input_with_deprecated_required.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class UseRequiredInput < Actor
+class DisallowNilOnInputWithDeprecatedRequired < Actor
   input :name, type: 'String', required: true
 
   def call; end